Generate the .designer.cs file too
This means when you reference the nuget the only thing you need to
do is set the 'Generator' property of the resx file to be
'MSBuild::GenerateStringResources'. After that the .Designer.cs and
the .Strings.cs will be generated automatically.

We additionally support the 'CustomToolNamespace' and 'Public' properties
to control the namespace and public/internal of the generated class.

A full file may look something like this:

<EmbeddedResource Update="Files\Resource.resx">
  <CustomToolNamespace>My.Alternate.Namespace</CustomToolNamespace>
  <Generator>MSBuild::GenerateStringResources</Generator>
  <Public>true</Public>
  <TargetClassName>MyCustomClass</TargetClassName>
</EmbeddedResource>

This will generate Resource.Designer.cs and Resource.Strings.cs.

The class in the '.Designer.cs' file will be called 'Resource' and
in the '.Strings.cs' file it is overrridden to be 'MyCustomClass'